Refactor: 주문완료 페이지 통합
Refs: #7
